Update log levels
